+++ admin/all.php
@@ -1,4 +1,5 @@
 <?php
 require_once 'hoo-save.php';
 require_once 'hoo-metabox.php';
-require_once 'hoo-template.php';
+require_once 'hoo-handlebars.php';
+require_once 'hoo-settings.php';