+++ assets/js/hoo.js
@@ -6,8 +6,16 @@ $(document).on('page:load', start);
 $(window).load(startOnLoad);
 
 function start() {
-  wooCheckout.init();
-  wooThankyou.init();
+  var $body = $('body');
+
+  if( $body.hasClass('woocommerce-checkout') ) {
+    wooCheckout.init();
+    wooThankyou.init();
+  }
+
+  if( $body.hasClass('single-product') ) {
+    wooSingle.init();
+  }
 }
 
 // functions that needs to run only after everything loads
@@ -15,6 +23,30 @@ function startOnLoad() {
   wooCheckout.moveEmailField();
 }
 
+///// SINGLE PRODUCT page
+
+var wooSingle = {
+  init: function() {
+    var $form = $('form.variations_form');
+    $form.on( 'found_variation', this.onFoundVariation );
+    $form.on( 'reset_data', this.onClear );
+  },
+
+  /*
+    After finished choosing all selection
+  */
+  onFoundVariation: function( e, variation ) {
+    $(this).siblings('.price').hide();
+  },
+
+  /*
+    After pressing "Clear" to remove all variant selection
+  */
+  onClear: function( e ) {
+    $(this).siblings('.price').show();
+  },
+
+};
 
 ///// CHECKOUT Page /////
 
@@ -24,7 +56,7 @@ var wooCheckout = {
 
     // add select wrapper
     $(document).on('country_to_state_changed', self._onCountryChanged.bind(self) );
-    
+
     // add active state to field
     $('.form-row input').each(this._checkActiveField);
     $('.form-row').on('change', 'input, select, textarea', this._checkActiveField);
@@ -75,9 +107,10 @@ var wooCheckout = {
     Move the error message inside the wrapper to maintain Flex layout
   */
   _moveErrorMessages: function(e) {
+    console.log( 'on error' );
     var $notice = $('.woocommerce-NoticeGroup');
 
-    $notice.insertAfter($('.entry-breadcrumbs') );
+    $notice.prependTo( $('.column-main') );
   }
 };