+++ public/hoo-init.php
@@ -2,18 +2,15 @@
 /*
   Generic function OR those that can't be run in template_redirect action.
 */
-class Hoo_General {
+class Hoo_Init {
 
   function __construct() {
     // Order Review
-    add_filter('woocommerce_cart_item_name', array($this, 'modify_cart_item_name'), 10, 3);
-    add_filter('woocommerce_order_item_name', array($this, 'modify_cart_item_name'), 10, 3);
+    add_filter('woocommerce_cart_item_name', array($this, 'add_thumbnail_to_order_table'), 10, 3);
+    add_filter('woocommerce_order_item_name', array($this, 'add_thumbnail_to_order_table'), 10, 3);
 
     // Templating
     add_filter('woocommerce_locate_template', array($this, 'woocommerce_locate_template'), 1, 3);
-
-    // Hide HELP tab
-    add_filter('woocommerce_enable_admin_help_tab', '__return_false');
   }
 
 
@@ -22,7 +19,7 @@ class Hoo_General {
 
     source: https://wisdmlabs.com/blog/override-woocommerce-templates-plugin/
   */
-  function woocommerce_locate_template($template, $template_name, $template_path) {
+  function woocommerce_locate_template( $template, $template_name, $template_path ) {
     global $woocommerce;
     $_template = $template;
     if(!$template_path) {
@@ -48,7 +45,7 @@ class Hoo_General {
   }
 
   /*
-    Add thumbnail to Order Review table
+    Add thumbnail to CART and ORDER table
     @filter woocommerce_cart_item_name
 
     @param $name (str)
@@ -57,9 +54,11 @@ class Hoo_General {
 
     @return str - The HTML tag for item name
   */
-  function modify_cart_item_name($name, $cart_item, $cart_item_key) {
-    if(is_checkout() ) {
-      $image = get_the_post_thumbnail($cart_item['product_id']);
+  function add_thumbnail_to_order_table( $name, $cart_item, $cart_item_key ) {
+    if( is_checkout() || is_wc_endpoint_url( 'view-order' )  ) {
+      $image = get_the_post_thumbnail( $cart_item['product_id'], 'thumbnail' );
+      $image = $image ? $image : sprintf("<img src='%s'>", wc_placeholder_img_src() );
+
       return $image . $name;
     } else {
       return $name;